+++ api/axios/requestHandler/othersManual/getOthersManual.api.tsx
@@ -1,21 +1,7 @@
-import { getRequest, postRequest } from '@/api/axios/common.apis'
+import { getRequest } from '@/api/axios/common.apis'
 
 export interface IKeyword {
   [key: string]: number
-}
-export interface Answer {
-  id: number
-  answer: string
-}
-
-export interface FormData {
-  data: {
-    nickname: string
-    answers: Answer[]
-    otherKeywordPercents: IKeyword
-  }
-  userType: string
-  userId: string
 }
 
 export type Item = {
@@ -50,17 +36,14 @@ export interface GetData {
 
 /* page get 요청 */
 export const fetchOthersManual = async () => {
-  const response: GetData = await getRequest('api/jff/others-manual?userId=1')
-
-  return response
-}
-
-/* page post 요청 */
-export const postOthersManual = async (data: FormData) => {
-  const response = await postRequest<string>(
-    '/api/jff/others-manual?userId=1',
-    data.data,
-  )
-
-  return response
+  try {
+    const response = await getRequest<GetData>('api/jff/others-manual?userId=1')
+    if (!response) {
+      throw new Error('no response')
+    }
+    return response
+  } catch (error) {
+    console.error(error)
+    throw error
+  }
 }
