+++ components/makeOthersManual/makeOthersManual.module.scss
@@ -100,10 +100,6 @@
   background-color: rgb(0 0 0 / 50%);
 }
 
-.modal {
-  max-height: 600px;
-}
-
 .formBoxWrapper {
   display: flex;
   flex-direction: column;
